Added :
* The DiPngImage class as a container for PNG image bytes. The data is usually compressed and 'decompressing' will NOT be supported by DiLib.
* 'Signature check' static methods inside di::image::internal::DiImageInternal for :
  - PNG file signature(8 bytes)
  - BMP file signature(2 bytes)

Updated :

* di::image::IGenericImage
* readme.md

+++ DiFileInternal.h
@@ -11,7 +11,7 @@ namespace di
 	{
 		namespace internal
 		{
-			using BinaryDataPair = std::pair<int, std::unique_ptr<uint8_t []>>;
+			using BinaryDataPair = std::pair<uint64_t, std::unique_ptr<uint8_t[]>>;
 
 			//Generic data storage class for file data
 			struct DiFileData
